requirements.txt
pydantic-settings == 2.7.1
python-dotenv == 1.0.1
python-binance == 1.0.27
pydantic == 2.11.0a1
pandas == 2.2.3
pandas-ta == 0.3.14b
plotly-express == 0.4.1
backtesting == 0.6.1
ta-lib == 0.6.0
